requirements.txt
numpy
pandas
streamlit
openai
requests
python-dotenv
azure-storage-blob 
azure-identity